+++ Firmware/ReFilFW/Thermistors.h
@@ -1,6 +1,6 @@
 
-#define THERMISTOR_0_PIN A7
-#define THERMISTOR_1_PIN A7 //TODO
+#define THERMISTOR_0_PIN A13
+#define THERMISTOR_1_PIN A14
 #define THERMISTOR_2_PIN A7 //TODO
 #define THERMISTOR_3_PIN A7 //TODO
